# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Review App API",
    "version": "1.0.0",
    "description": "API documentation for the Software Design Artifacts Review Application",
    "contact": {
      "name": "API Support",
      "email": "support@reviewapp.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "schemas": {
      "RubricCriteria": {
        "type": "object",
        "required": [
          "name",
          "score"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the criteria"
          },
          "description": {
            "type": "string",
            "description": "Description of the criteria"
          },
          "score": {
            "type": "number",
            "minimum": 0,
            "maximum": 5,
            "description": "Score for the criteria"
          },
          "comment": {
            "type": "string",
            "description": "Optional comment for the criteria"
          }
        }
      },
      "RubricEvaluation": {
        "type": "object",
        "required": [
          "rubricType",
          "project",
          "evaluator"
        ],
        "properties": {
          "rubricType": {
            "type": "string",
            "enum": [
              "requirements",
              "stories",
              "activityDiagrams",
              "useCaseDiagrams",
              "sequenceDiagrams",
              "classDiagrams",
              "designPatterns",
              "mockups"
            ]
          },
          "criteria": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RubricCriteria"
            }
          },
          "overallScore": {
            "type": "number",
            "minimum": 0,
            "maximum": 5
          },
          "generalComment": {
            "type": "string"
          },
          "project": {
            "type": "string",
            "description": "Project ID"
          },
          "evaluator": {
            "type": "string",
            "description": "Evaluator ID"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/rubric-evaluations/{projectId}/{rubricType}": {
      "get": {
        "summary": "Get rubric evaluation for a project",
        "tags": [
          "Rubric Evaluations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "rubricType",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "requirements",
                "stories",
                "activityDiagrams",
                "useCaseDiagrams",
                "sequenceDiagrams",
                "classDiagrams",
                "designPatterns",
                "mockups"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved rubric evaluation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RubricEvaluation"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/rubric-evaluations": {
      "post": {
        "summary": "Save a complete rubric evaluation",
        "tags": [
          "Rubric Evaluations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RubricEvaluation"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Evaluation saved"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/rubric-evaluations/general-comment": {
      "post": {
        "summary": "Save a general comment on a rubric evaluation",
        "tags": [
          "Rubric Evaluations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "projectId",
                  "rubricType",
                  "comment"
                ],
                "properties": {
                  "projectId": {
                    "type": "string"
                  },
                  "rubricType": {
                    "type": "string",
                    "enum": [
                      "requirements",
                      "stories",
                      "activityDiagrams",
                      "useCaseDiagrams",
                      "sequenceDiagrams",
                      "classDiagrams",
                      "designPatterns",
                      "mockups"
                    ]
                  },
                  "comment": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Comment saved"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/rubric-evaluations/{projectId}/{rubricType}/calculate-score": {
      "get": {
        "summary": "Calculate overall score for a rubric evaluation",
        "tags": [
          "Rubric Evaluations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "rubricType",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "requirements",
                "stories",
                "activityDiagrams",
                "useCaseDiagrams",
                "sequenceDiagrams",
                "classDiagrams",
                "designPatterns",
                "mockups"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Calculated score",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "overallScore": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/stories/project/{projectId}": {
      "get": {
        "summary": "Get all stories for a project",
        "tags": [
          "Stories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Project ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of stories"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Not authorized to access this project"
          }
        }
      }
    },
    "/api/stories/{id}": {
      "get": {
        "summary": "Get a single story",
        "tags": [
          "Stories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Story ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Story details with linked requirements"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Not authorized to view this story"
          },
          "404": {
            "description": "Story not found"
          }
        }
      },
      "put": {
        "summary": "Update a story",
        "tags": [
          "Stories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Story ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "text": {
                    "type": "string",
                    "description": "Gherkin code"
                  },
                  "requirementsLinked": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Array of requirement IDs"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Story updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Not authorized to update this story"
          },
          "404": {
            "description": "Story not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a story",
        "tags": [
          "Stories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Story ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Story deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Not authorized to delete this story"
          },
          "404": {
            "description": "Story not found"
          }
        }
      }
    },
    "/api/stories": {
      "post": {
        "summary": "Create a new story",
        "tags": [
          "Stories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "text",
                  "projectId"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "text": {
                    "type": "string",
                    "description": "Gherkin code"
                  },
                  "requirementsLinked": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Array of requirement IDs"
                  },
                  "projectId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Story created successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Not authorized to add stories to this project"
          }
        }
      }
    },
...
  },
...
}